HW1/service2/requirements.txt
blinker==1.8.2
boto3==1.35.39
botocore==1.35.39
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
huggingface-hub==0.25.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.1
numpy==2.0.2
packaging==24.1
pika==1.3.2
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
s3transfer==0.10.3
safetensors==0.4.5
six==1.16.0
tokenizers==0.20.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
Werkzeug==3.0.4
